added boardArray to replace jquery selectors

+++ src/components/game/Game.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useMemo, useRef, useCallback, createContext } from 'react';
+import { useState, useEffect, useMemo, useRef, useCallback, createContext, memo } from 'react';
 import classes from './Board.module.css';
 import { BoardMemo } from './Board';
 import $ from 'jquery';
@@ -12,10 +12,11 @@ import { Container, Row, Col } from 'react-bootstrap';
 
 export const PlayerContext = createContext();
 export const TurnContext = createContext({ turn: ['white', 'pe2', 'pe4'], setTurn: () => {} });
+export const BoardContext = createContext();
 export const EnPassentContext = createContext();
 export const CheckContext = createContext();
 
-export default function Game(props) {
+export default memo(function Game(props) {
   // default FEN notation: rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq -
   // after e4: rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq e3
   // e3 is the en passent square, the numbers tell you how many consecutive squares are emtpy from left to right from the perspective of white
@@ -28,7 +29,7 @@ export default function Game(props) {
     });
   });
 
-  const board = useRef([
+  const [boardArray, setBoardArray] = useState([
     ['r', 'n', 'b', 'q', 'k', 'b', 'n', 'r'],
     ['p', 'p', 'p', 'p', 'p', 'p', 'p', 'p'], // black
     ['1', '1', '1', '1', '1', '1', '1', '1'],
@@ -54,30 +55,35 @@ export default function Game(props) {
 
   const setUpTurnChange = useRef(false);
   const fixBoardArray = useCallback((FEN) => {
-    // takes FEN as an argument, and fixes the board.current which we use as a middleman between a move, 'pe2, pe4' to putting that in FEN
+    // takes FEN as an argument, and fixes the board which we use as a middleman between a move, 'pe2, pe4' to putting that in FEN
     // 'rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq e3'
     let FEN_index = 0;
+    const tempboard = [];
     for (let i = 0; i < 8; i++) {
+      const row = [];
       for (let j = 0; j < 8; j++, FEN_index++) {
         if (FEN[FEN_index] === '/') FEN_index++;
         let num = parseInt(FEN[FEN_index]);
         if (num) {
           if (num === 1) {
-            board.current[i][j] = '1';
+            row.push('1');
           }
           else {
-            while (num > 1) {
-              board.current[i][j] = '1';
+            while (num > 0) {
+              row.push('1');
               num--;
               j++;
             }
+            j--;
           }
         }
         else {
-          board.current[i][j] = FEN[FEN_index];
+          row.push(FEN[FEN_index]);
         }
       }
+      tempboard.push(row);
     }
+    setBoardArray(tempboard);
   }, []);
   const fixTurnFromFEN = useCallback((FEN) => {
     // switching from b in FEN to w immediately after a white move, in both local storage and in firebase
@@ -192,39 +198,38 @@ export default function Game(props) {
   );
 
   const setLastMoveFromOtherUser = useRef(true);
+  const setFENFromOtherUser = useRef(true);
   useEffect(
     () => {
       // if turn is 'white' or 'black' then this is the wrong user/socket as it has no newLocation
-      if (turn.length === 5) {
-        return;
-      }
+      if (turn.length === 5) return;
 
       let newLocation = [turn[1], turn[2]];
       setLastMoveFromOtherUser.current = false;
+      setFENFromOtherUser.current = false;
 
       let column = newLocation[0].charCodeAt(1) - 97; // gets e from pe2 and converts that to 4th column (3 in array)
       let row = parseInt(newLocation[0][2]) - 1; // gets 2 from pe2 and converts that to the 2nd column (1 in array)
-      let piece = board.current[7 - row][column];
-      board.current[7 - row][column] = '1'; // prev square is now empty
 
+      let piece = boardArray[7 - row][column];
+      const tempBoard = boardArray;
+      tempBoard[7 - row][column] = '1'; // prev square is now empty
       let newColumn = newLocation[1].charCodeAt(1) - 97;
       let newRow = parseInt(newLocation[1][2]) - 1;
-      board.current[7 - newRow][newColumn] = piece;
+      tempBoard[7 - newRow][newColumn] = piece;
 
       // counts pieces from left to right, counting up 1's inbetween pieces to get new FEN
       let temp_FEN = '';
       for (let i = 0; i < 8; i++) {
         let spaces = 0;
         for (let j = 0; j < 8; j++) {
-          if (board.current[i][j] === '1') {
-            spaces++;
-          }
+          if (tempBoard[i][j] === '1') spaces++;
           else {
             if (spaces > 0) {
               temp_FEN += spaces;
               spaces = 0;
             }
-            temp_FEN += board.current[i][j];
+            temp_FEN += tempBoard[i][j];
           }
         }
         if (spaces > 0) {
@@ -246,6 +251,7 @@ export default function Game(props) {
       }
       else temp_FEN += ' -';
 
+      setBoardArray(tempBoard);
       setFEN(temp_FEN);
       localStorage.setItem('FEN', temp_FEN);
       update(ref(database, 'Games/' + gameID.current), {
@@ -262,8 +268,12 @@ export default function Game(props) {
           check: null
         });
       }
+
+      if (playerColor === 'white') setTurn('black');
+      else setTurn('white');
     },
-    [fixBoardArray, turn]
+    //eslint-disable-next-line
+    [turn]
   );
 
   // these onValue events are for sockets listening that did not trigger the event.
@@ -273,12 +283,15 @@ export default function Game(props) {
     const dbRef = ref(database, 'Games/' + gameID.current + '/FEN');
     onValue(dbRef, (snapshot) => {
       // if snapshot.val() === FEN that means that the user who updated the FEN in the first place is running this
-      if (snapshot.exists() && snapshot.val() !== FEN) {
+      if (snapshot.exists() && setFENFromOtherUser.current === true && snapshot.val() !== FEN) {
         const temp_FEN = snapshot.val();
         fixBoardArray(temp_FEN);
         localStorage.setItem('FEN', temp_FEN);
         setFEN(temp_FEN);
         off(dbRef);
+      }
+      else if (snapshot.exists() && setFENFromOtherUser.current === false) {
+        setFENFromOtherUser.current = true;
       }
     });
   });
@@ -287,15 +300,16 @@ export default function Game(props) {
   useEffect(() => {
     const dbRef = ref(database, 'Games/' + gameID.current + '/check');
     onValue(dbRef, (snapshot) => {
-      if (snapshot.exists()) {
-        if (
-          ((snapshot.val()[0][0] === snapshot.val()[0][0].toLowerCase() && playerColor === 'white') ||
-            (snapshot.val()[0][0] === snapshot.val()[0][0].toUpperCase() && playerColor === 'black')) &&
-          (check === null || JSON.stringify(snapshot.val()) === JSON.stringify(check))
-        ) {
-          setCheck(snapshot.val());
-          off(dbRef);
-        }
+      if (!snapshot.exists()) return;
+      console.log(JSON.stringify(snapshot.val()), JSON.stringify(check));
+      let checkingPiece = snapshot.val()[0][0];
+      if (
+        ((checkingPiece === checkingPiece.toLowerCase() && playerColor === 'white') ||
+          (checkingPiece === checkingPiece.toUpperCase() && playerColor === 'black')) &&
+        (check === null || JSON.stringify(snapshot.val()) !== JSON.stringify(check))
+      ) {
+        setCheck(snapshot.val());
+        off(dbRef);
       }
     });
   });
@@ -303,26 +317,25 @@ export default function Game(props) {
   useEffect(() => {
     const dbRef = ref(database, 'Games/' + gameID.current + '/lastMove');
     onValue(dbRef, (snapshot) => {
-      let newLocation = [turn[1], turn[2]];
-      if (snapshot.exists()) {
-        if (setLastMoveFromOtherUser.current === false) setLastMoveFromOtherUser.current = true;
-        else {
-          newLocation = snapshot.val();
-          let old_location = $('#' + newLocation[0]);
-          if (old_location.length === 0) return;
+      if (!snapshot.exists()) return;
+      else if (setLastMoveFromOtherUser.current === false) setLastMoveFromOtherUser.current = true;
+      else {
+        let newLocation = snapshot.val();
+        let oldLocation = $('#' + newLocation[0]);
+        if (oldLocation.length === 0) return;
 
-          let destination = $('[id*=' + newLocation[1][1] + newLocation[1][2])[0];
-          let capturedPiece = $('#' + destination.id)[0].firstChild;
-          if (capturedPiece) $('#' + capturedPiece.id).remove();
-
-          old_location.appendTo($('#S' + newLocation[1][1] + newLocation[1][2]));
-          old_location.attr('id', old_location[0].id[0] + newLocation[1][1] + newLocation[1][2]);
-
-          if (playerColor === 'white') setTurn('white');
-          else setTurn('black');
+        let destination = boardArray[newLocation[1].charCodeAt(1) - 'a'.charCodeAt(0)][8 - parseInt(newLocation[1][2])];
+        if (destination !== '1') {
+          $('#' + destination + newLocation[1][1] + newLocation[1][2]).remove();
         }
-        off(dbRef);
+
+        oldLocation.appendTo($('#S' + newLocation[1][1] + newLocation[1][2]));
+        oldLocation.attr('id', oldLocation[0].id[0] + newLocation[1][1] + newLocation[1][2]);
+
+        if (playerColor === 'white') setTurn('white');
+        else setTurn('black');
       }
+      off(dbRef);
     });
   });
 
@@ -351,13 +364,15 @@ export default function Game(props) {
 
             <Col>
               <CapturedPanel>
-                <TurnContext.Provider value={turnValue}>
-                  <EnPassentContext.Provider value={enPassentSquare}>
-                    <CheckContext.Provider value={check}>
-                      <BoardMemo currentUser={currentUserID.current} FEN={FEN} />
-                    </CheckContext.Provider>
-                  </EnPassentContext.Provider>
-                </TurnContext.Provider>
+                <BoardContext.Provider value={boardArray}>
+                  <TurnContext.Provider value={turnValue}>
+                    <EnPassentContext.Provider value={enPassentSquare}>
+                      <CheckContext.Provider value={check}>
+                        <BoardMemo currentUser={currentUserID.current} FEN={FEN} />
+                      </CheckContext.Provider>
+                    </EnPassentContext.Provider>
+                  </TurnContext.Provider>
+                </BoardContext.Provider>
               </CapturedPanel>
             </Col>
           </PlayerContext.Provider>
@@ -369,4 +384,4 @@ export default function Game(props) {
       </Container>
     </div>
   );
-}
+});
